feat(service): support cookies
