+++ packages/platform-web/package.json
@@ -85,7 +85,6 @@
     "@shuvi/toolpack": "1.0.0",
     "@shuvi/utils": "1.0.0",
     "content-type": "1.0.4",
-    "cookie": "0.4.1",
     "core-js": "3.6.5",
     "ejs": "3.1.5",
     "fs-extra": "9.0.1",
@@ -105,7 +104,6 @@
   "devDependencies": {
     "@testing-library/react": "^13.2.0",
     "@types/content-type": "^1.1.5",
-    "@types/cookie": "^0.4.1",
     "@types/ejs": "^3.1.0",
     "@types/raw-body": "^2.3.0",
     "@types/react": "18.0.9",
