+++ packages/service/package.json
@@ -43,6 +43,7 @@
     "fs-extra": "9.0.1",
     "http-proxy-middleware": "1.0.6",
     "rimraf": "3.0.2",
+    "cookie": "0.4.1",
     "send": "0.17.1",
     "ws": "8.2.3",
     "dotenv": "16.0.3",
@@ -55,6 +56,7 @@
     "@types/babel__traverse": "^7.14.2",
     "@types/cross-spawn": "6.0.1",
     "@types/detect-port": "1.3.0",
+    "@types/cookie": "^0.4.1",
     "@types/ejs": "3.0.4",
     "@types/ws": "8.2.0"
   }
